Initial version
Still very raw; this setup for creating new Ubuntu 18 Virtualbox images was created some time ago...
It was thrown together from bits and pieces of information online.
It is absurdly complex what is going on, but it does work.

It is hardcoded to a specific version of Ubuntu; it may or may not work with newer versions.
Good luck making it do anything else except exactly what it was originally setup to do.

There are a bunch of manual steps that must be done even after running go.pl, before booting up the VM.
See the comments within go.pl